+++ stubs/synapse/synapse_rust/__init__.pyi
